+++ packages/http-client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-cloud-sdk/http-client
 
+## 2.7.0
+
+### Patch Changes
+
+- Updated dependencies [3bff42e1]
+- Updated dependencies [010a46fa]
+  - @sap-cloud-sdk/connectivity@2.7.0
+  - @sap-cloud-sdk/util@2.7.0
+
 ## 2.6.0
 
 ### Minor Changes