+++ packages/mail-client/CHANGELOG.md
@@ -1,1 +1,10 @@
 # @sap-cloud-sdk/mail-client
+
+## 2.7.0
+
+### Patch Changes
+
+- Updated dependencies [3bff42e1]
+- Updated dependencies [010a46fa]
+  - @sap-cloud-sdk/connectivity@2.7.0
+  - @sap-cloud-sdk/util@2.7.0