+++ test-packages/integration-tests/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-cloud-sdk/integration-tests
 
+## 2.7.0
+
+### Patch Changes
+
+- Updated dependencies [3bff42e1]
+- Updated dependencies [010a46fa]
+  - @sap-cloud-sdk/connectivity@2.7.0
+  - @sap-cloud-sdk/mail-client@2.7.0
+  - @sap-cloud-sdk/test-util@2.7.0
+  - @sap-cloud-sdk/util@2.7.0
+  - @sap-cloud-sdk/test-services-odata-v2@2.7.0
+  - @sap-cloud-sdk/test-services-odata-v4@2.7.0
+
 ## 2.6.0
 
 ### Patch Changes